requirements.txt
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
binance-connector==1.17.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==37.0.4
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
requests==2.28.1
service-identity==21.1.0
six==1.16.0
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
urllib3==1.26.12
zope.interface==5.4.0
